+++ samples/Demo/Beef.Demo.Business/DataSvc/RobotDataSvc.cs
@@ -1,0 +1,7 @@
+﻿namespace Beef.Demo.Business.DataSvc
+{
+    public partial class RobotDataSvc
+    {
+        private Task<Result<Robot?>> GetOnImplementationAsync(Guid id) => _data.GetAsync(id);
+    }
+}